% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d3_modules}
\alias{d3_modules}
\title{D3 module sizes.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{size}{Number of lines in d3 modules}
  \item{path}{path to d3 module, separated by slash}
  ...
}}
\source{
\url{https://d3js.org/}
}
\usage{
d3_modules
}
\description{
A dataset containing the size and path of d3 modules.
}
\keyword{datasets}
