% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiervis-package.R
\docType{package}
\name{hiervis-package}
\alias{hiervis-package}
\title{hiervis: Interactive Hierarchical Visualizations for R/Shiny}
\description{
The foo package provides several types of hierarchical 
visualizations, most of which are interactive. The
visualizations include Sankey, Partition, Sunburst and Icicle.
}
\details{
All visualizations can be accessed with the function hiervis,
and explored with the function hiervis_demo, which starts a 
Shiny app.
}
