% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiervis.R
\name{hiervis}
\alias{hiervis}
\title{Create a hierarchical visualization}
\usage{
hiervis(data, vis, width = NULL, height = NULL, elementId = NULL,
  nameField = "name", valueField = "value", pathSep = NULL,
  parentField = NULL, stat = "count")
}
\arguments{
\item{data}{data.frame containing}

\item{vis}{One of "sankey", "sunburst", "partition", "treemap"}

\item{width}{Width of SVG element.}

\item{height}{Height of SVG element.}

\item{pathSep}{If pathSep is supplied, consider the nameField to be a path.}
}
\description{
Creates hierarchical visualizations based on data frame.
}
\examples{
data(Titanic)

## Tabular data does not need any extra arguments
hiervis(Titanic, "sankey")
hiervis(HairEyeColor, "vertical sankey")

## For tabular data (data.frames) with a path, supply nameField, pathSep and valueField
hiervis(d3_modules, "sunburst", nameField = "path", pathSep = "/", valueField = "size")

## For tabular data (data.frames) with parent-child information, supply nameField and parentField
data <- data.frame(name = c("Root Node", "Node A", "Node B", "Leaf Node A.1", "Leaf Node A.2"), parent = c(NA, "Root Node", "Root Node", "Node A", "Node A"))
hiervis(data, "sankey", nameField = "name", parentField = "parent", stat = "count")
}
