% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiervis-demo.R
\name{hiervis_demo}
\alias{hiervis_demo}
\title{Shiny demo for hiervis}
\usage{
hiervis_demo(data = NULL, ..., title = "hiervis demo")
}
\arguments{
\item{data}{Custom data}

\item{...}{Additional arguments to hiervis()}

\item{title}{Title of the Shiny page}
}
\value{
Shiny app
}
\description{
Shiny demo for hiervis
}
\examples{
\donttest{hiervis_demo()}
}
